+++ aws/ecs-application/security_group.tf
@@ -2,14 +2,14 @@
 aws_security_group provides a security group resource for ECS service.
 https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/security_group.html
 */
-resource "aws_security_group" "main" {
+resource "aws_security_group" "this" {
   vpc_id      = var.vpc_id
   name        = "${var.name}-sg"
   description = "Allow inbound HTTP traffic to service from ALB only"
 
 
   ingress {
-    description     = "allow from ALB only"
+    description     = "Allow from ALB only"
     from_port       = var.container_port
     to_port         = var.container_port
     protocol        = "tcp"